Refactoring done for isuues #28, #29, #30, #31, #32, #33, #34 and #35 tickets.
